+++ elevenlabs_rs/src/endpoints/convai/widget.rs
@@ -126,19 +126,19 @@ impl CreateWidgetAvatarBody {
 }
 
 impl TryFrom<&CreateWidgetAvatarBody> for RequestBody {
-    type Error = Box<dyn std::error::Error + Send + Sync>;
+    type Error = crate::error::Error;
 
     fn try_from(body: &CreateWidgetAvatarBody) -> Result<Self> {
         let path = Path::new(&body.avatar_file);
         let bytes = std::fs::read(path)?;
         let mut part = Part::bytes(bytes);
-        let file_path_str = path.to_str().ok_or(Box::new(Error::PathNotValidUTF8))?;
+        let file_path_str = path.to_str().ok_or(Error::PathNotValidUTF8)?;
         part = part.file_name(file_path_str.to_string());
         let mime_subtype = path
             .extension()
-            .ok_or(Box::new(Error::FileExtensionNotFound))?
+            .ok_or(Error::FileExtensionNotFound)?
             .to_str()
-            .ok_or(Box::new(Error::FileExtensionNotValidUTF8))?;
+            .ok_or(Error::FileExtensionNotValidUTF8)?;
         let mime = format!("image/{}", mime_subtype);
         part = part.mime_str(&mime)?;
         Ok(RequestBody::Multipart(