+++ elevenlabs_rs/src/lib.rs
@@ -32,7 +32,7 @@
 //!     Ok(())
 //! }
 //! ```
-pub use crate::client::{ElevenLabsClient, Result};
+pub use crate::client::{ApiResponse, ElevenLabsClient, Result};
 pub use crate::shared::query_params::*;
 pub use crate::shared::{DefaultVoice, Language, LegacyVoice, Model, VoiceSettings};
 pub use bytes::Bytes;